reconfigured the directory to have a separate server folder with sub-directories
