+++ server/controllers/marketController.js
@@ -1,8 +1,7 @@
-var Market = require('./farmModel.js');
+var Market = require('../model/farmModel.js');
 var Q = require('q');
 var rp = require('request-promise');
-// import {GOOGLE_API_KEY} from './API_KEYS.js';
-var GOOGLE_API_KEY = require('./API_KEYS');
+var MarketQuery = require('../methods/marketMethods');
 
 var getAllFarms = Q.nbind(Market.find, Market);
 
@@ -12,7 +11,7 @@ var replaceSpaceInAddress = (address) => {
 
 module.exports = {
 
-	allMarkets: function(req,res,next){
+	allMarkets: function(req,res, next){
 		console.log("allMarkets....")
 		getAllFarms({})
 		  .then(function(farms){
@@ -24,20 +23,20 @@ module.exports = {
 	},
 
 	getLocationMarkets: (req, res, next) => {
-		var address = replaceSpaceInAddress(req.body);
+		var address = replaceSpaceInAddress('1216 Broadway New York, NY');
+		// var address =
 		rp.get(
 			`https://maps.googleapis.com/maps/api/geocode/json?address=${address}`
 		).then((data) => {
-			
-			console.log('successfully got geocode', data);
+			var coordinates = JSON.parse(data).results[0].geometry.location;
+			// console.log('successfully got geocode', lat);
+			MarketQuery.fetchMarkets(coordinates);
 		}).catch((err) => {
 			console.error('Failed', err);
 		});
 		// some function to get a list of farms that match the DB query
 
 
-
-		// convert the `req.url` and `req.body` (which will be the radius with the address as center) into a longitutude/latitude
 		// send the query result to front-end as res.json()
 	}
 };
