+++ server/methods/marketMethods.js
@@ -1,0 +1,3 @@
+var mongoose = require('mongoose');
+var Market = require('../model/farmModel.js');
+var Q = require('q');
