+++ server/server.js
@@ -1,6 +1,6 @@
 var express = require('express');
 var bodyParser = require('body-parser');
-var routes = require('./routes.js');
+var routes = require('./routes/routes.js');
 
 var mongoose = require('mongoose');
 mongoose.connect('mongodb://user1:user1@ds117899.mlab.com:17899/fazzar');
